Small fixes

+++ nd4j/nd4j-backends/nd4j-tests/src/test/java/org/nd4j/autodiff/opvalidation/MiscOpValidation.java
@@ -796,9 +796,9 @@ public class MiscOpValidation extends BaseOpValidation {
 
         INDArray ia = Nd4j.trueVector(new double[]{2,2}).castTo(DataType.INT);
         double value = 42;
-        INDArray out = Nd4j.create(DataType.DOUBLE, 2,2);
+        INDArray out = Nd4j.create(DataType.FLOAT, 2,2);
         OpTestCase op = new OpTestCase(new Fill(ia, out, value));
-        INDArray expOut = Nd4j.valueArrayOf(new int[]{2,2}, 42.0);
+        INDArray expOut = Nd4j.valueArrayOf(new long[]{2,2}, 42.0f);
 
         op.expectedOutput(0, expOut);
         String err = OpValidation.validate(op);
